+++ src/refactoring/pages/AdminPage.tsx
@@ -1,16 +1,25 @@
 import { useState } from "react";
-import { Coupon, Discount, Product } from "../../types";
+import { Coupon, Product } from "../../types";
 import { useProductContext, useCouponContext } from "../contexts";
 
 export const AdminPage = () => {
-  const { products, updateProduct, addProduct } = useProductContext();
+  const {
+    products,
+    openProductIds,
+    editingProduct,
+    newDiscount,
+    addProduct,
+    toggleProductAccordion,
+    editProduct,
+    productNameUpdate,
+    priceUpdate,
+    editComplete,
+    stockUpdate,
+    addDiscount,
+    removeDiscount,
+  } = useProductContext();
   const { coupons, addCoupon } = useCouponContext();
-  const [openProductIds, setOpenProductIds] = useState<Set<string>>(new Set());
-  const [editingProduct, setEditingProduct] = useState<Product | null>(null);
-  const [newDiscount, setNewDiscount] = useState<Discount>({
-    quantity: 0,
-    rate: 0,
-  });
+
   const [newCoupon, setNewCoupon] = useState<Coupon>({
     name: "",
     code: "",
@@ -24,81 +33,6 @@ export const AdminPage = () => {
     stock: 0,
     discounts: [],
   });
-
-  const toggleProductAccordion = (productId: string) => {
-    setOpenProductIds((prev) => {
-      const newSet = new Set(prev);
-      if (newSet.has(productId)) {
-        newSet.delete(productId);
-      } else {
-        newSet.add(productId);
-      }
-      return newSet;
-    });
-  };
-
-  // handleEditProduct 함수 수정
-  const handleEditProduct = (product: Product) => {
-    setEditingProduct({ ...product });
-  };
-
-  // 새로운 핸들러 함수 추가
-  const handleProductNameUpdate = (productId: string, newName: string) => {
-    if (editingProduct && editingProduct.id === productId) {
-      const updatedProduct = { ...editingProduct, name: newName };
-      setEditingProduct(updatedProduct);
-    }
-  };
-
-  // 새로운 핸들러 함수 추가
-  const handlePriceUpdate = (productId: string, newPrice: number) => {
-    if (editingProduct && editingProduct.id === productId) {
-      const updatedProduct = { ...editingProduct, price: newPrice };
-      setEditingProduct(updatedProduct);
-    }
-  };
-
-  // 수정 완료 핸들러 함수 추가
-  const handleEditComplete = () => {
-    if (editingProduct) {
-      updateProduct(editingProduct);
-      setEditingProduct(null);
-    }
-  };
-
-  const handleStockUpdate = (productId: string, newStock: number) => {
-    const updatedProduct = products.find((p) => p.id === productId);
-    if (updatedProduct) {
-      const newProduct = { ...updatedProduct, stock: newStock };
-      updateProduct(newProduct);
-      setEditingProduct(newProduct);
-    }
-  };
-
-  const handleAddDiscount = (productId: string) => {
-    const updatedProduct = products.find((p) => p.id === productId);
-    if (updatedProduct && editingProduct) {
-      const newProduct = {
-        ...updatedProduct,
-        discounts: [...updatedProduct.discounts, newDiscount],
-      };
-      updateProduct(newProduct);
-      setEditingProduct(newProduct);
-      setNewDiscount({ quantity: 0, rate: 0 });
-    }
-  };
-
-  const handleRemoveDiscount = (productId: string, index: number) => {
-    const updatedProduct = products.find((p) => p.id === productId);
-    if (updatedProduct) {
-      const newProduct = {
-        ...updatedProduct,
-        discounts: updatedProduct.discounts.filter((_, i) => i !== index),
-      };
-      updateProduct(newProduct);
-      setEditingProduct(newProduct);
-    }
-  };
 
   const handleAddCoupon = () => {
     addCoupon(newCoupon);
@@ -226,10 +160,7 @@ export const AdminPage = () => {
                             type="text"
                             value={editingProduct.name}
                             onChange={(e) =>
-                              handleProductNameUpdate(
-                                product.id,
-                                e.target.value
-                              )
+                              productNameUpdate(product.id, e.target.value)
                             }
                             className="w-full p-2 border rounded"
                           />
@@ -240,10 +171,7 @@ export const AdminPage = () => {
                             type="number"
                             value={editingProduct.price}
                             onChange={(e) =>
-                              handlePriceUpdate(
-                                product.id,
-                                parseInt(e.target.value)
-                              )
+                              priceUpdate(product.id, parseInt(e.target.value))
                             }
                             className="w-full p-2 border rounded"
                           />
@@ -254,10 +182,7 @@ export const AdminPage = () => {
                             type="number"
                             value={editingProduct.stock}
                             onChange={(e) =>
-                              handleStockUpdate(
-                                product.id,
-                                parseInt(e.target.value)
-                              )
+                              stockUpdate(product.id, parseInt(e.target.value))
                             }
                             className="w-full p-2 border rounded"
                           />
@@ -278,7 +203,7 @@ export const AdminPage = () => {
                               </span>
                               <button
                                 onClick={() =>
-                                  handleRemoveDiscount(product.id, index)
+                                  removeDiscount(product.id, index)
                                 }
                                 className="bg-red-500 text-white px-2 py-1 rounded hover:bg-red-600"
                               >
@@ -291,28 +216,28 @@ export const AdminPage = () => {
                               type="number"
                               placeholder="수량"
                               value={newDiscount.quantity}
-                              onChange={(e) =>
-                                setNewDiscount({
-                                  ...newDiscount,
-                                  quantity: parseInt(e.target.value),
-                                })
-                              }
+                              // onChange={(e) =>
+                              //   setNewDiscount({
+                              //     ...newDiscount,
+                              //     quantity: parseInt(e.target.value),
+                              //   })
+                              // }
                               className="w-1/3 p-2 border rounded"
                             />
                             <input
                               type="number"
                               placeholder="할인율 (%)"
                               value={newDiscount.rate * 100}
-                              onChange={(e) =>
-                                setNewDiscount({
-                                  ...newDiscount,
-                                  rate: parseInt(e.target.value) / 100,
-                                })
-                              }
+                              // onChange={(e) =>
+                              //   setNewDiscount({
+                              //     ...newDiscount,
+                              //     rate: parseInt(e.target.value) / 100,
+                              //   })
+                              // }
                               className="w-1/3 p-2 border rounded"
                             />
                             <button
-                              onClick={() => handleAddDiscount(product.id)}
+                              onClick={() => addDiscount(product.id)}
                               className="w-1/3 bg-blue-500 text-white p-2 rounded hover:bg-blue-600"
                             >
                               할인 추가
@@ -320,7 +245,7 @@ export const AdminPage = () => {
                           </div>
                         </div>
                         <button
-                          onClick={handleEditComplete}
+                          onClick={editComplete}
                           className="bg-green-500 text-white px-2 py-1 rounded hover:bg-green-600 mt-2"
                         >
                           수정 완료
@@ -338,7 +263,7 @@ export const AdminPage = () => {
                         ))}
                         <button
                           data-testid="modify-button"
-                          onClick={() => handleEditProduct(product)}
+                          onClick={() => editProduct(product)}
                           className="bg-blue-500 text-white px-2 py-1 rounded hover:bg-blue-600 mt-2"
                         >
                           수정
